formatted

+++ kg_bacdive/transform_utils/ontology/ontology_transform.py
@@ -1,4 +1,6 @@
 """Ontology transform module."""
+import gzip
+import shutil
 from pathlib import Path
 from typing import Optional, Union
 
@@ -8,14 +10,12 @@ from kgx.cli.cli_utils import transform
 from kg_bacdive.utils.robot_utils import convert_to_json
 
 from ..transform import Transform
-import gzip
-import shutil
 
 ONTOLOGIES = {
     # "HpTransform": "hp.json",
     # 'GoTransform': 'go-plus.json',
-    "NCBITransform": "ncbitaxon.json",
-    'ChebiTransform': 'chebi.owl.gz',
+    "ncbitaxon": "ncbitaxon.json",
+    "chebi": "chebi.owl.gz",
     # "EnvoTransform": "envo.json",
     # 'GoTransform': 'go.json'
 }
@@ -57,15 +57,17 @@ class OntologyTransform(Transform):
         :return: None.
         """
         if data_file.suffixes == [".owl", ".gz"]:
-            # Unzip the file
-            print(f"Decompressing {data_file}...")
-            with gzip.open(data_file, 'rb') as f_in:
-                with open(data_file.parent / data_file.stem, 'wb') as f_out:
-                    shutil.copyfileobj(f_in, f_out)
+            json_path = str(data_file).replace("owl.gz", "json")
+            if not Path(json_path).is_file():
+                # Unzip the file
+                print(f"Decompressing {data_file}...")
+                with gzip.open(data_file, "rb") as f_in:
+                    with open(data_file.parent / data_file.stem, "wb") as f_out:
+                        shutil.copyfileobj(f_in, f_out)
 
-            print(f"Converting {data_file} to obojson...")
-            convert_to_json(str(self.input_base_dir), name)
-            data_file = str(data_file).replace("owl.gz", "json")
+                print(f"Converting {data_file} to obojson...")
+                convert_to_json(str(self.input_base_dir), name)
+            data_file = json_path
 
         transform(
             inputs=[data_file],
